pyproject.toml
[project]
name = "maeva-prospect-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat==0.5.1",
    "autogen-ext[openai]==0.5.1",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
